- bug fixed (position pid gain & velocity pid gain sync write). - added velocity_to_value_ratio to DXL Pro-H series.

+++ robotis_device/include/robotis_device/dynamixel_state.h
@@ -59,12 +59,12 @@ public:
   double    goal_position_;
   double    goal_velocity_;
   double    goal_torque_;
-  double    position_p_gain_;
-  double    position_i_gain_;
-  double    position_d_gain_;
-  double    velocity_p_gain_;
-  double    velocity_i_gain_;
-  double    velocity_d_gain_;
+  int       position_p_gain_;
+  int       position_i_gain_;
+  int       position_d_gain_;
+  int       velocity_p_gain_;
+  int       velocity_i_gain_;
+  int       velocity_d_gain_;
 
   std::map<std::string, uint32_t> bulk_read_table_;
 
@@ -78,13 +78,13 @@ public:
       goal_position_(0.0),
       goal_velocity_(0.0),
       goal_torque_(0.0),
-      position_p_gain_(0.0),
-      position_i_gain_(0.0),
-      position_d_gain_(0.0),
-      velocity_p_gain_(0.0),
-      velocity_i_gain_(0.0),
-      velocity_d_gain_(0.0),
-      position_offset_(0.0)
+      position_p_gain_(0),
+      position_i_gain_(0),
+      position_d_gain_(0),
+      velocity_p_gain_(0),
+      velocity_i_gain_(0),
+      velocity_d_gain_(0),
+      position_offset_(0)
   {
     bulk_read_table_.clear();
   }
